refactor: remove unused parser and unify types
