+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { PodcastProvider } from "@/context/PodcastContext";
-import Header from "@/components/Header"; // <-- Add this line
+import Header from "@/components/Header";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -16,7 +16,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     <html lang="en">
       <body className={`${inter.className} bg-gray-50 text-gray-900`}>
         <PodcastProvider>
-          <Header /> {/* Add the header above content */}
+          <Header />
           <div className="min-h-screen">{children}</div>
         </PodcastProvider>
       </body>
