+++ src/lib/types.ts
@@ -1,0 +1,11 @@
+export type Episode = {
+  title: string;
+  pubDate: string;
+};
+
+export type PodcastFeed = {
+  title: string;
+  description: string;
+  episodes: Episode[];
+  url: string;
+};
